Modernize Putchar to PrintRune.
1. char => Rune
Because `char` is a C type, not a Go one (although it
has `byte` for an ASCII character). `rune` is the prefered Go type
for a single glyph.

Also, when iterating on a string, it will decompose it with runes, not
bytes.

2. Put => Print
Because Go already use the verb "Print" to output text on console.